chore: make use of drop_all parameter
