+++ projects/pgai/tests/vectorizer/extensions/fixtures/alembic/autogen_env.py.template
@@ -7,7 +7,7 @@ from sqlalchemy import engine_from_config, pool
 
 from pgai.alembic.autogenerate import compare_vectorizers
 from pgai.alembic.operations import CreateVectorizerOp, DropVectorizerOp
-from pgai.alembic.operations import render_create_vectorizer, render_drop_vectorizer
+from pgai.alembic.autogenerate import render_create_vectorizer, render_drop_vectorizer
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
